Tensorflow/3_text_bert/1.requirements.txt
absl-py==0.9.0
asn1crypto==0.24.0
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
cryptography==2.1.4
cycler==0.10.0
dataclasses==0.7
filelock==3.0.12
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.6
joblib==0.16.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
Markdown==3.1.1
matplotlib==3.2.2
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
packaging==20.4
pandas==1.0.5
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==2.4.7
python-apt==1.6.4
python-dateutil==2.8.1
pytz==2020.1
pyxdg==0.25
PyYAML==5.3.1
regex==2020.7.14
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
sacremoses==0.0.43
scipy==1.4.1
SecretStorage==2.3.1
sentencepiece==0.1.91
six==1.13.0
tensorboard==2.1.0
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0
termcolor==1.1.0
tokenizers==0.8.1rc1
tqdm==4.48.0
transformers==3.0.2
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
xlrd==1.2.0
